sample-files: introduce local-arago64-sam67 config sample file
Compared to the origin local-arago64-v2, DL_DIR and SSTATE_DIR are no
longer set. This allows us to pass these directories from build host
environment with the ability to re-use downloads and sstate-cache
directories on the build host thereby saving disk space and compile
time.

In case that DL_DIR and SSTATE_DIR are not passed from environment,
bitbake will fall back to the default ${TOPDIR}/downloads and
${TOPDIR}/sstate-cache directories.

Note that re-use of these directories is safe, as bitbake tracks any
configuration parameters for download and build and will rebuild a
package if there is a mismatch.

Signed-off-by: Thomas Schaefer <thomas.schaefer@kontron.com>